refactor(integration-tests-o11y): minimize cfg attributes
